# Docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "quant_api",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/events": {
      "get": {
        "tags": [],
        "operationId": "events",
        "responses": {
          "200": {
            "description": "SSE completed artifact snapshots; no job launcher"
          }
        }
      }
    },
    "/api/models": {
      "get": {
        "tags": [],
        "operationId": "models",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModelArtifact"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/models/{artifact_id}": {
      "get": {
        "tags": [],
        "operationId": "model",
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelArtifact"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs": {
      "get": {
        "tags": [],
        "operationId": "runs",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunManifest"
                  }
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/manifest": {
      "get": {
        "tags": [],
        "operationId": "manifest",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunManifest"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/{artifact}": {
      "get": {
        "tags": [],
        "operationId": "artifact",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "artifact",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "max_points",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "asof",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "instrument",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Arrow IPC or JSON artifact"
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Instrument": {
        "type": "object",
        "required": [
          "id",
          "kind"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          }
        }
      },
      "ModelArtifact": {
        "type": "object",
        "required": [
          "artifact_id",
          "metadata",
          "onnx_present"
        ],
        "properties": {
          "artifact_id": {
            "type": "string"
          },
          "metadata": {},
          "onnx_present": {
            "type": "boolean"
          },
          "training_log": {},
          "validation": {}
        }
      },
      "Provenance": {
        "type": "object",
        "required": [
          "git_commit",
          "config_hash",
          "data_version",
          "model_artifact_id",
          "source"
        ],
        "properties": {
          "config_hash": {
            "type": "string"
          },
          "data_version": {
            "type": "string"
          },
          "git_commit": {
            "type": "string"
          },
          "model_artifact_id": {
            "type": "string"
          },
          "source": {
            "type": "string"
          }
        }
      },
      "RunManifest": {
        "type": "object",
        "required": [
          "schema_version",
          "run_id",
          "created_at",
          "kind",
          "split",
          "provenance",
          "instruments",
          "capabilities",
          "artifacts",
          "metrics",
          "warnings"
        ],
        "properties": {
          "artifacts": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "capabilities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "created_at": {
            "type": "string"
          },
          "instruments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Instrument"
            }
          },
          "kind": {
            "type": "string"
          },
          "metrics": {
            "type": "object",
            "additionalProperties": {
              "type": "number",
              "format": "double"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "run_id": {
            "type": "string"
          },
          "schema_version": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "split": {
            "type": "string",
            "description": "test, validation, train, or unverified. A label alone is not proof of OOS."
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}